Move http exceptions to errors module.
