+++ src/backend/aspen/app/app.py
@@ -18,7 +18,7 @@ from flask_cors import CORS
 from sentry_sdk.integrations.flask import FlaskIntegration
 from sqlalchemy.orm.exc import NoResultFound
 
-from aspen.app import exceptions as ex
+from aspen.error import http_exceptions as ex
 from aspen.app.aspen_app import AspenApp
 from aspen.app.views.api_utils import get_usergroup_query
 from aspen.config.config import Config
@@ -54,6 +54,10 @@ application = AspenApp(
     static_folder=str(static_folder),
     aspen_config=aspen_config,
 )
+
+@application.errorhandler(ex.AspenException)
+def handle_bad_request(err):
+    return err.make_response()
 
 allowed_origins = []
 frontend_url = os.getenv("FRONTEND_URL")
@@ -179,18 +183,15 @@ def requires_auth(f):
         # Redirect to Login page
         if not auth0_user_id:
             return redirect("/login")
-        try:
-            with session_scope(application.DATABASE_INTERFACE) as db_session:
-                g.db_session = db_session
-                found_auth_user = setup_userinfo(auth0_user_id)
-                if not found_auth_user:
-                    # login attempt from user not in DB
-                    # TODO: return response to user that they do not have an account with aspen.
-                    return redirect("/login")
-                else:
-                    g.auth_user = found_auth_user
-                return f(*args, **kwargs)
-        except ex.AspenException as err:
-            return err.make_response()
+        with session_scope(application.DATABASE_INTERFACE) as db_session:
+            g.db_session = db_session
+            found_auth_user = setup_userinfo(auth0_user_id)
+            if not found_auth_user:
+                # login attempt from user not in DB
+                # TODO: return response to user that they do not have an account with aspen.
+                return redirect("/login")
+            else:
+                g.auth_user = found_auth_user
+            return f(*args, **kwargs)
 
     return decorated
