+++ src/backend/aspen/app/views/sample.py
@@ -10,7 +10,7 @@ from flask import g, jsonify, request, Response, stream_with_context
 from sqlalchemy import and_, or_
 from sqlalchemy.orm import joinedload
 
-from aspen.app import exceptions as ex
+from aspen.error import http_exceptions as ex
 from aspen.app.app import application, requires_auth
 from aspen.app.views import api_utils
 from aspen.app.views.api_utils import check_valid_sequence
