+++ src/backend/aspen/error/http_exceptions.py
